+++ src/main/java/kitchenpos/deliveryorders/infra/DefaultKitchenridersClient.java
@@ -1,9 +1,11 @@
-package kitchenpos.infra;
+package kitchenpos.deliveryorders.infra;
 
 import org.springframework.stereotype.Component;
 
 import java.math.BigDecimal;
 import java.util.UUID;
+
+import kitchenpos.deliveryorders.domain.KitchenridersClient;
 
 @Component
 public class DefaultKitchenridersClient implements KitchenridersClient {
